chore: move integration capability code out of graphql/mutations/helpers
- pushEstimateTo*/create*Task (+test) → integrations/<svc>/; importers use ./
- RemoteRepoIntegration: per-service repo types live with their integration
- select.ts: derive TeamMemberIntegrationAuth meta narrowing from IntegrationAuthMetaByService
