+++ packages/server/integrations/github/GitHubRepo.ts
@@ -1,0 +1,5 @@
+export type GitHubRepo = {
+  id: string
+  nameWithOwner: string
+  service: 'github'
+}
