+++ packages/server/integrations/github/GitHubServerManager.ts
@@ -3,13 +3,13 @@ import type {GraphQLResolveInfo} from 'graphql'
 import GitHubIssueId from '../../../client/shared/gqlIds/GitHubIssueId'
 import GitHubRepoId from '../../../client/shared/gqlIds/GitHubRepoId'
 import type {GQLContext} from '../../graphql/graphql'
-import createGitHubTask from '../../graphql/mutations/helpers/createGitHubTask'
 import type {GitHubAuth} from '../../postgres/types'
 import type {AddCommentMutation, AddCommentMutationVariables} from '../../types/githubTypes'
 import getGitHubRequest from '../../utils/getGitHubRequest'
 import addComment from '../../utils/githubQueries/addComment.graphql'
 import makeCreateGitHubTaskComment from '../../utils/makeCreateGitHubTaskComment'
 import type {CreateTaskResponse, TaskIntegrationManager} from '../platform/TaskIntegrationManager'
+import createGitHubTask from './createGitHubTask'
 
 export default class GitHubServerManager implements TaskIntegrationManager {
   public title = 'GitHub'
