+++ packages/server/integrations/github/createGitHubTask.ts
@@ -1,18 +1,18 @@
 import type {JSONContent} from '@tiptap/core'
 import type {GraphQLResolveInfo} from 'graphql'
 import {splitTipTapContent} from 'parabol-client/shared/tiptap/splitTipTapContent'
-import type {GitHubAuth} from '../../../postgres/types'
+import type {GQLContext} from '../../graphql/graphql'
+import type {GitHubAuth} from '../../postgres/types'
 import type {
   CreateIssueMutation,
   CreateIssueMutationVariables,
   GetRepoInfoQuery,
   GetRepoInfoQueryVariables
-} from '../../../types/githubTypes'
-import {convertTipTapToMarkdown} from '../../../utils/convertTipTapToMarkdown'
-import getGitHubRequest from '../../../utils/getGitHubRequest'
-import createIssueMutation from '../../../utils/githubQueries/createIssue.graphql'
-import getRepoInfo from '../../../utils/githubQueries/getRepoInfo.graphql'
-import type {GQLContext} from '../../graphql'
+} from '../../types/githubTypes'
+import {convertTipTapToMarkdown} from '../../utils/convertTipTapToMarkdown'
+import getGitHubRequest from '../../utils/getGitHubRequest'
+import createIssueMutation from '../../utils/githubQueries/createIssue.graphql'
+import getRepoInfo from '../../utils/githubQueries/getRepoInfo.graphql'
 
 const createGitHubTask = async (
   rawContent: JSONContent,
