+++ packages/server/integrations/gitlab/GitLabProject.ts
@@ -1,0 +1,6 @@
+export type GitLabProject = {
+  id: string
+  service: 'gitlab'
+  __typename: 'Project'
+  fullPath: string
+}
