+++ packages/server/integrations/jira/createJiraTask.ts
@@ -2,8 +2,8 @@ import type {JSONContent} from '@tiptap/react'
 import {convertTiptapToADF} from 'parabol-client/shared/tiptap/convertTipTapToADF'
 import {splitTipTapContent} from 'parabol-client/shared/tiptap/splitTipTapContent'
 import {RateLimitError} from 'parabol-client/utils/AtlassianManager'
-import type {AtlassianAuth} from '../../../postgres/types'
-import AtlassianServerManager from '../../../utils/AtlassianServerManager'
+import type {AtlassianAuth} from '../../postgres/types'
+import AtlassianServerManager from '../../utils/AtlassianServerManager'
 
 const createJiraTask = async (
   rawContent: JSONContent,
