+++ packages/server/integrations/platform/RemoteRepoIntegration.ts
@@ -1,28 +1,9 @@
-import type {LinearRepoIntegration} from 'parabol-client/shared/gqlIds/IntegrationRepoId'
 import type {JiraGQLProject} from '../../dataloader/atlassianLoaders'
 import type {AzureAccountProject} from '../../dataloader/azureDevOpsLoaders'
 import type {JiraServerProject} from '../../dataloader/jiraServerLoaders'
-
-type GitHubRepo = {
-  id: string
-  nameWithOwner: string
-  service: 'github'
-}
-
-type GitLabProject = {
-  id: string
-  service: 'gitlab'
-  __typename: 'Project'
-  fullPath: string
-}
-
-type LinearProjectRepo = LinearRepoIntegration & {
-  displayName: string
-}
-
-type LinearTeamRepo = LinearRepoIntegration & {
-  name: string
-}
+import type {GitHubRepo} from '../github/GitHubRepo'
+import type {GitLabProject} from '../gitlab/GitLabProject'
+import type {LinearRepo} from '../linear/LinearRepo'
 
 export type RemoteRepoIntegration =
   | JiraGQLProject
@@ -30,5 +11,4 @@ export type RemoteRepoIntegration =
   | GitLabProject
   | JiraServerProject
   | AzureAccountProject
-  | LinearProjectRepo
-  | LinearTeamRepo
+  | LinearRepo
